change spyOn to fn for mocking window.scrollTo
jest.spyOn still gives a not implemented error when trying to spy on
window.scrollTo because that is not implemented in jest dom. directly
mocking window.scrollTo creates a mock at that method name rather than trying
to find the existing method at that name and hijacking it with a mock.